Refactor and reorganize diagnostics logic
